path fix?

+++ src/motions.py
@@ -1,5 +1,6 @@
 import gzip
 import os
+import sys
 
 
 # generally just ess up, but also considered adjusting
@@ -55,33 +56,48 @@ def ess_up_adjust_noncached(angle):
         index += 1
 
 
+# from https://stackoverflow.com/a/70405825
+def get_data_folder():
+    # path of your data in same folder of main .py or added using --add-data
+    if getattr(sys, 'frozen', False):
+        data_folder_path = sys._MEIPASS
+    else:
+        data_folder_path = os.path.dirname(
+            os.path.abspath(sys.modules['__main__'].__file__)
+        )
+    return data_folder_path
+
+
 CAMERA_SNAPS = []
-snapPath = os.path.dirname(os.path.abspath(__file__)) + "/../res/camera_snaps.txt"
-if not os.path.isfile(snapPath):
-    snapPath = snapPath.replace("/..", "")
-
 try:
-    with gzip.open(f"{snapPath}.gz", "rt") as cam:
-        for line in cam:
-            if line.strip() == "False":
-                CAMERA_SNAPS.append(False)
-            else:
-                CAMERA_SNAPS.append(int(line))
-except:
-    camera_angles = []
-    with open(snapPath, "r") as f:
-        for line in f:
-            camera_angles.append(int(line.strip(), 16))
+    snapPath = f"{get_data_folder()}/../res/camera_snaps.txt"
+    if not os.path.isfile(snapPath):
+        snapPath = snapPath.replace("/..", "")
 
-    for angle in range(0xFFFF + 1):
-        if (angle % 0x1000) == 0:
-            print(f"Caching camera movements ({hex(angle)})...", end="\r")
-        CAMERA_SNAPS.append(ess_up_adjust_noncached(angle))
-    print("\nDone.")
+    if not os.path.isfile(f"{snapPath}.gz"):
+        camera_angles = []
+        with open(snapPath, "r") as f:
+            for line in f:
+                camera_angles.append(int(line.strip(), 16))
 
-    with gzip.open(f"{snapPath}.gz", "wt") as cam:
-        for angle in CAMERA_SNAPS:
-            print(angle, file=cam)
+        for angle in range(0xFFFF + 1):
+            if (angle % 0x1000) == 0:
+                print(f"Caching camera movements ({hex(angle)})...", end="\r")
+            CAMERA_SNAPS.append(ess_up_adjust_noncached(angle))
+        print("\nDone.")
+
+        with gzip.open(f"{snapPath}.gz", "wt") as cam:
+            for angle in CAMERA_SNAPS:
+                print(angle, file=cam)
+    else:
+        with gzip.open(f"{snapPath}.gz", "rt") as cam:
+            for line in cam:
+                if line.strip() == "False":
+                    CAMERA_SNAPS.append(False)
+                else:
+                    CAMERA_SNAPS.append(int(line))
+except Exception as e:
+    raise print(e)
 
 
 def ess_up_adjust(angle):
